kernel call into rust-space
This commit brings the kernel into Rust-space. We add Cargo to manage
building the Rust code and dependencies. Rust requires a few last
changes to boot.asm and linker.ld, but it's more than worth the effort.

Some comments indicated that enabling SSE makes working with interrupts
difficult and that a better approach is to tell Rust (read: LLVM) not to
use SSE-based instructions. So we may need to revisit these changes in
the future.