initial support for vga writer
This commit introduces a Writer that supports printing text to the
screen via the VGA buffer. The writer is meant to be owned (Unique), but
this is not yet enforced. Text will wrap at 80 columns. Functionality
exists to clear the screen, and to scroll it when filled, but these are
untested because the stack is still too small to include a large string.